[REFACTOR] [#28] Renamed the position layout to base layout. [ADDED] Documentation for the base activity.
Fixes #28
