+++ app/src/main/java/com/hossainkhan/android/demo/browse/MainActivity.kt
@@ -19,13 +19,11 @@ package com.hossainkhan.android.demo.browse
 import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
 import android.support.v7.widget.GridLayoutManager
-import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
-import android.view.View
 import com.hossainkhan.android.demo.R
 import com.hossainkhan.android.demo.data.AppDataStore
-import com.hossainkhan.android.demo.layoutpositioning.LayoutPositioningDemoActivity
-import com.hossainkhan.android.demo.layoutpositioning.LayoutVisibilityGoneActivity
+import com.hossainkhan.android.demo.layoutpreview.LayoutPreviewBaseActivity
+import com.hossainkhan.android.demo.layoutpreview.LayoutVisibilityGoneActivity
 import dagger.android.AndroidInjection
 import timber.log.Timber
 import javax.inject.Inject
@@ -73,7 +71,7 @@ class MainActivity : AppCompatActivity() {
                         .createStartIntent(this))
             }
             else -> {
-                startActivity(LayoutPositioningDemoActivity
+                startActivity(LayoutPreviewBaseActivity
                         .createStartIntent(this, layoutResId))
             }
         }
