+++ app/src/main/res/layout/preview_positioning_top_left.xml
@@ -21,7 +21,7 @@
     android:layout_height="match_parent"
     android:foreground="@drawable/ic_grid_overlay"
     android:foregroundTint="@color/grid_tint"
-    tools:context=".layoutpositioning.LayoutPositioningDemoActivity">
+    tools:context=".layoutpreview.LayoutPreviewBaseActivity">
 
 
     <!--
